[Behat] Extract ProductTranslationsTrait from product pages
